Fix PeMS cache self-heal crash + ARIMA/Prophet missing train_start bound + stale ridership dtype docs; carry forward PR #110 (2026-09-23 daily audit)

+++ evaluation/validators.py
@@ -90,7 +90,7 @@ REQUIRED_COLUMNS = [
 EXPECTED_DTYPES = {
     "timestamp":  "datetime64[ns, America/Los_Angeles]",
     "station_id": "object",
-    "ridership":  "float64",
+    "ridership":  "int64",
 }
